se agrega el patron repositorio al model DealBranding, se separan los modelos de negocio en los 3 principales

+++ database/migrations/2023_05_20_003719_create_deals_branding_table.php
@@ -33,7 +33,7 @@ return new class extends Migration {
       $table->tinyInteger('available_budget');
       $table->tinyInteger('saving_profit_percent');
       $table->boolean('client_report')->default(0);
-      $table->integer('external_operation');
+      $table->tinyInteger('external_operation');
       $table->timestamps();
     });
   }
